Add files via upload
In this comprehensive Python project, I implemented a decision tree classifier to predict flight delays using flight data from Interpretation Challenges I. The decision tree, visualized graphically with a depth of 3 and presented in text form, provides a clear understanding of the classification process. Moving on to Explainable Boosting Machines (EBM), the advantages and disadvantages are explored, shedding light on its interpretability challenges. Subsequently, an EBM model is trained using a subset of features and data. The feature importance and interactions are analyzed globally, and the ROC curve is plotted to explain the model's overall performance. Additionally, Scope Rules are visualized and explained, offering insights into the decision-making process. Lastly, a special dataset is used to explore feature importance and interactions locally, providing a nuanced understanding of the model's behavior. This project, available on GitHub, serves as a valuable resource for those interested in interpretability challenges in machine learning, offering both code and comprehensive explanations for each step of the analysis.